update & bugs fix:
bug: added variable in sfmake.inc containing path to scifor_completion script to simplify
construction of completion call in Makefile.
bug: found in sciforvars.sh script, "add_library_to_system *$*$LIB" double $-symbol in bash variable definition. fixed.
bug: in functions_bethe a problem appear in the gfbether routine, Hilbert transform of the bethe lattice on real-axis, because of
possible 1/0.0 case is w=0.0. Fixed by adding a small shift which prevents floating overflow.

update: in the process of updating the FFT module we introduced a better version of the DERIVATE module, which contains now
new routines to evaluate the 1,2,3,4,nth derivative using higher order formula (default=4, admitted are: [1 for n=1,2], 2, 4, 6)