update: changed the configuration files: bin/sciforvars.sh, etc/library.conf and include/sfmake.inc
The scope of the change is to put SciFor on diet.
The package rely on third-party library to work, such as FFTPACK, MINPACK, QUADPACK, LAPACK, BLAS,
and MKL (not mandatory btw).
However, these packages are not required to compile the SciFor routines but are only necessary at
compilation time.

Thus I decided to leave at the user will to add these packages to the system and to tune the
compilation arguments (i.e. the list of libraries to be linked against to) in the specific Makefile,
i.e. the complicated argument list has been removed from sfmake.inc file which contains  the -lscifor
or -lscifor_deb argument plus specification of INCLUDE path for SciFor modules.

Indeed the SciFor library is added to the system when sourcing sciforvars.sh (just like INTEL/MKL).
In principle this ensures that .mod files are in the CPATH/FPATH environment variable, so that no additional
-I<SFROOT/include> statement is necessary, and that library is in the (LD_)LIBRARY_PATH, so that
no further -L<SFROOT/lib> statement is necessary (again just like INTEL/MKL).

In practice in order to keep using the DEBUG channel, the INCLUDE statement must be retained (and not implicitly
assumed in the F/CPATH).

This will make sligtly harder the life of the programmer, but leave he/she free to decide what to include or not.
Free will is hard to life with, but guarantees freedom and happiness.