Version 2.0 : structure of the SciFor library is entirely changed.
 - new directories structure. Main directory contains a *_install* dir which
contains the SciFor package to be installed. The install.sh script is in charge
of configuring and compiling on the fly the library and put wrap everything in a
specific directory locally. This latter (named with the platform specified at
installation time) contains the lib, include and bin directories to added to system.

 - normal and debug installation have separete route now, so there is no need for
all the complications introduced in previous 1.x versions.

 - an environment module is configured for SciFor (intel and intel_debug). This can be
found in the package at https://github.com/aamaricci/EnvModules.git .
In the future one can update install.sh script to print out the corresponding module
at installation time. For the time being the destination directory contains a suitable
bash script which can be used to add a library (i.e. a dir containing lib, include, man, etc.)
to the system.

 - the library still rely on MKL and other third party libraries like FFTPACK, MINPACK, etc...
this is esplictly states in the environment module configuration when loading SciFor as pre-requisites.
MKL can be get rid of by updating FFT modules, using the soon to be finished new FFT versions.
Environment module can be used to satisfy dependencies. Otherwise the bash_add_lib.sh script
can become handy. The third party packages can all be found at the following sites:
https://github.com/aamaricci/FFTpack.git
https://github.com/aamaricci/Minpack.git
https://github.com/aamaricci/Quadpack.git
https://github.com/aamaricci/Blas.git
https://github.com/aamaricci/Lapack.git

Alternatively a full installer of all (and beyond) libraries can be cloned from:
https://github.com/aamaricci/INSTALL_Libs.git

- All SciFor linking and including variable have been removed. This is in spirit the most important
change for two reasons. 1) If library is correctly added to the system compilation will automatically
look for the modules and libraries in the correct places, so there is no need to specify location.
Moreover having debug version on a separate channel simplify this task at the cost of reinitialize
the environment each time, which can be easily done using environment module. 2) The user is free to
add or remove other libraries when needed at compilation time in the specific Makefile.

This will make sligtly harder the life of the programmer, but leave he/she free to decide what to include or not.
Free will is hard to live with, but guarantees more freedom and thus more happiness.

- in this version GNU platform does not compile, both because of the MKL dependence (which can be got rid of
by manually including MKL header files) and other small issues. I will solve this problem in the next release.

- TIMER has been updated so to not use the declaration "carriagecontrol=fortran", which is heavily compiler
dependent.