NEW VERSION v3.0
This is a slightly new version of the library, diverging from version2.0
Some of the modules, specific or dedicated to DMFT or CondMat tasks, have been
moved to a new library named DMFT_TOOLS (tentative name).
We try to follow revision 0.7 of SciPy for this version of SciFor.

From now on SciFor will contain only general and generic modules.