New version based on FFTPACK and SciFor/FFT_FFTPACK for FFT routines in FFTGF. This version is independent of MKL (and compiles with gfortran).
ATTENTION: BY USING THIS VERSION MANY CODES OR PARTS OF CODES USING fftgf_ ROUTINES
WILL STOP WORKING. FIX AND TEST AGAIN.

update: fix installer and module file.

update2: M.Balestieri found a hidden but serious bug involving transforms in the near band
insulating case. To be fixed this required to update all the FFT transforms:

NEW FFTGF SCHEME:
the transforms now are all defined with proper tail subtraction, in both directions.
This was tested for the 1b Hubbard model at U=4 (Mott Insulator) and other tests for the
non-interacting solution with at or away from half-filling as well as for the band insulators.

The tail subtraction uses the simple scheme: C0 + C1/iw + C2/(iw)**2 + C2/(iw)**3 for the direct
transform (iw-->tau) or the other scheme (thanks a lot A.Valli for this):
1/(iw + a + b/iw) which reduces to the first scheme for some choice of a and b, but fives a nicer
tail in tau.

The routines are now divived in iw2tau, tau2iw and separated for Green's function and Sigma
functions.

Already tested in the ED code (in the driver because ED does not rely on FFTGF).