Version 0.3 This version had numerous bugs fixed (see log).
FFTGF is a generic interface with full tail treatment and extraction, tested in IPT and NEQ_DMFT
TIGHT_BINDING module added to the library. This can be used to construct tb hamiltonian. Much more
to do in this module in the future.
SQUARE_LATTICE was removed as this was only used in NEQ_DMFT that now has been rewritten completely
and can be used without this specific module.