version v0.4:
small improvement of FFTGF: the tail coefficients C in the subroutines call
have been made optional.
If present standard version of the tail is calculated.
If not, the tail is calculated starting from the INPUT behavior at large frequency, as
in the old version of the routines.
This apparently improves the calculations in some cases, e.g. G(k,iw) to obtain n(k)
distribution as tested (bug fixing) in the non-eq DMFT code.

update: added further options in build_model in TIGHT_BINDING
update: small changes in FFTGF.