version 0.5:
this version of the library has few changes but all
modules have been renamed to DMFT_xxx for better consistency and to
avoid conflict.

routines in DMFT_TOOLS module are now contained in DMFT_FFTAUX as these are
auxiliaty to the DMFT_FFTGF module

TOOLS has been renamed to DMFT_MISC to avoid conclicts.

The modules now use SCIFOR and load moduled of type SF_xxx

The library feature a template directory which now contain the linked lists
modules, which can be used as starting point to write the required one.