VERSION V0.5.1
update DMFT_TIGHT_BINDING.

-------------- WARNING ------------------
In this new release I changed the shapes
of Hk matrices so to have the dependence
on the number of k-points always in the
last dimension, rather than in the first
(as it was in the previous release).
-----------------------------------------

bug fixed:
+solve_HkR_along_BZpath had a bug following
the previous warning: the array of the eigen-
values Eval(Nlat,Norb,Nk) where printed
as Eval(Nk,Nlat,Norb).

update:
introduced pos_comment in DMFT_PARSE_INPUT
to set placement of the comment in the
input file. Default is now 46.