update:
- big change in the SF_INTEGRATE package.
routines are now divided in different files, depending on their scope:
 + integrate_func_1d/2d.f90: contains routines to integrate 1d/2d function
with explicit interface, using trapeziodal and simpson's rule.
 + integrate_sample_1d/2d.f90 : contains routines to integrate 1d/2d data arrays
using trapeziodal and simpson's rule.

Former routines are kept, but are superseeded by a new interface which accept
optional arguments and performs the same task as before depending on the given input.
This one should be used instead of the former routines, but this i) breaks compilation
of the codes, ii) should be tested first.

- ready to provide a simple interface to the main quadpack routines, this
will requires quadpack to be installed in the system.

PROPOSAL:
it would be nice to force the installation of local version of third party libraries
together with scifor, so that the user should only install scifor to have all the
main libraries available.
This can possibly include also arpack as soon as an interface to it is provided
and a primitive version of sparse_matrix and sparse_linalg is included.