v3.6 THIS IS A NEW AND RESHAPED VERSION OF SCIENTIFIC FORTRAN.
changes:
- in this version all the third party libraries have been integrated in the opt directory
including: P-Arpack, Blas, Dmft_Tools, FFTpack, LApack, MINpack, QUADpack.

- the libraries are compiled and placed all together in the same target directory.
A modules file and configvars script are used to load the full package of libraries.

This change enables to avoid installation of many 3rd party libraries keeping
everything independent. In this way installing SciFor becomes a way to have
a set of math libraries in the system for Fortran development.