UPDATE:
The installer was updated with a better handling of the prefix directory.
If prefix is not writable, admin passwd is asked to the user via the sudo -v command.
This ensure the possibility to write the output into an external directory (e.g. /opt/scifor/gnu)
and avoid to keep it inside the code directory.

A better handling of the cleaning strategy is required. For the time being one can rely
on re-running the command with the --clean option to delete the .o and .mod files stored
locally in the code directory. This decrease the amount of files to sync and to keep.
Indeed, once the 3rd party libraries are generated the installation reduces to an update
of the scifor part only.