v3.8.6 UPDATE
- code updated with a new version of the INPUT PARSER (SF_PARSE_INPUT.f90).
This new version is based on pointer and array of pointers which point directly
to the input variable as declared in the calling program (I know this is prone to memory leak).
In this way the LIST of INPUT variables is automatically updated as the user change
value to any of the input variables.
I OBSERVED SOME MINOR MISBEHAVIOR AFTER ALTERATION OF A VARIABLE VALUE:
in particular the value was reported as updated only after a print of the variable itself (that
I suppose corresponds to perform any operation on the variable, I am still investigating this).

This version of the INPUT PARSE provide a meaningful memory pool that can be used to setup
few significant variables that must be retrieved in a project and can not be passed in other way.

-- update:
removed add_input_variable and get_input_variable. The INPUT PARSE do just its job: parse the input variables.
I prepared another module that take care of creating a list of control variables to be used in DMFT_TOOLS
as a memory pool for some procedure.