v3.9 NEW
This is a "newer" version.
Structure of the code has been changed slightly, but integrating all the third party
libraries (blas,lapack,arpack,fftpack,minpack,quadpack) directly into the src of the
SciFortran library.
On one hand this improves a lot the self-consistency of the whole library
One the other it ease a lot the installation and compilation process.

The library has been switched to MPI, because it directly (see below) use the
MPI parallel scheme for some procedures.
From now on the library will be developed in a serial/parallel way using
pre-compilation.

UPDATES:
- Updated installer compatible with the new directory structure. Simpler than before

- MPI support. The installer checks existence of MPI installation and compatibility with
required platform (intel, gnu). MPI installation is default, serial version of the library
can be compiled with the --no-mpi option.

- NEW update script. To avoid problems with reinstallation of the whole 3rd party libraries
an update script is now provided. the update script checks existence of a SciFortran installation
and re-generate the SciFor library via recompilation.

- NEW interface to (P)Arpack library: SF_SP_LINALG.f90. This module provides an interface to the
Double/Complex main routines in Arpack as eigensolvers for sparse matrices. The interface follows
closely the SciPy one. A parallel version (parpack) is generated using precompilation flags and
MPI.