v4.1.0 NEW RELEASE
This version is built via a CMake.

The entire structure of the library has been adapted to the Cmake compilation
to ensure best portability and efficient compilation on all platform.
Perhaps some adjustments are in order. Tested against the ED code.

The compilation has a default prefix in the user home.
Prefix can be set using -DPREFIX=prefix

The compilation has a default USE_MPI flag (ON).
Non MPI is executed using -DUSE_MPI=OFF