v4.1.1 UPDATE
Few fixes to the new version CMakefile.txt:
- added *pkg-config* configuration file as an extra usage option.
The .pc file containing all information about the installed
library, is placed in the installation directory in etc.
It is pre-pended to the env variable PKG_CONFIG_PATH with the
environment module or within configvars.sh script. Otherwise
it would be enough to add the target directory containing the .pc
file to this env variable:
export PKG_CONFIG_PATH=path-to-dir-containing-pc-file

- fixed an issue with the Scifor Versions module

- CMakefile provides useful after-installation informations.

- Added post-installation script to add env.module and pkg-config files
into user home (avoiding permissions problems, this step should
not be run as sudo)