v4.1.4 UPDATE:
- The *str* procedure, transforming any implicit variable (dble,cmplx,chr,bool) into a
string of minimal length has been updated introducing the possibility of fixing the
number of digits for dble and cmplx variables: str(r8,d=digits) or str(c,d=digits).
update 1: solved a tiny but nasty bug related to the log estimate of the order of
magnitude in str. if r8 or c are zero log10 gets w_=infty and code tries to allocate a
an infinite character(len=w_). solved using if condition.
update1: fixed a small bug here.

- All 3rd parties libraries have a LICENSE file containing the official license with which
the library was published. This should have been done before. Note that all libraries are
either Public Domain or GPL or BSD. While SciFortran is LGPL (less generic GPL).