v4.3.1a BUG FIX
Fixed a bug in CMakeLists.txt for the update of the library.

Bug: the version_sha1.inc file was generated at the CMAKE call.
So updates of the library without recompilation did not re-generate
this file ending up in a wrong description of the actual version
to be used. Now the file is generated at compilation time (make).

the _sha1 file is removed after installation ([sudo] make install).