v4.3.3 UPDATE:
We analyzed the CG in the Num. Rec. implementation (fmin_cg).
We discovered circumstances for which the gradient is nearly or exactly
zero. This leads to a misbehavior of MNBRACK in (D)LINMIN, which
tries to bracket the minima along one direction (that of the gradient).
Being zero this bracketing can not be resolved and caused MNBRACK to
increase its interval to NaN (infty).

When this happens we added a stop condition (isnan --> exit).
We added a further input parameter (optionl) in fmin_cg to control
the quality of the gradient evaluation when this is not passed by the
user. default is zero (as before).