v4.3.4 BUG FIX:
Fixed a bug in *str* in IOFILE.
The transformation of a real/cmplx to a string was affected by a bug.
The trigger was realized by passing a real number with of an order of
magnitude so small to compensate the required decimals (default d=6).
This required the code to allocate a string of negative value which in
some compilers triggered a "allocating too much memory" error.

Fixed by doing the only reasonable thing: copy the Fortran default.
Looking at print*, fortran print out a decimal of an order of magnitude
smaller than -1 already in ESw.d format.
> str does the same now.
The price to pay is that this print occupy a width slight larger (4ch) with
respect to an equivalent one with order of magnitude >= -1. Too bad...