v4.4.0 UPDATE [BREAK SOME BACK COMPATIBILITY]
- SF_MPI: removed get_MPI_Q/R/Chunk
these routines should not be used anymore. The parallel procedures in SciFor are meant to
be independent of the way one decides to split the work.

>> THIS UPDATE POTENTIALLY BREAK SOME BACKCOMPATIBLITY  <<
- SF_SP_LINALG: major update in arpack_mpi (sp_eigh) and lanczos_mpi (sp_lanczos_eigh, tridiag).
In this update these sparse diagonalization routines perform the work independently of the MPI split
of the incoming vector. The user must be sure that MPI split is the same between the calls of the
HxV operator.
Previously there was an implicit assumption on the split of the vectors to be used in the matrix-vector
product. This has been removed: the routines have as input/output already the splitted vectors.
It is on the user to split/recollect them before and after the use.