v4.4.8 UPDATE
- SF_SP_LINALG: updated arpack interface and fixed minor (but possibly annoying) bugs.
There is an unsolved question concerning what to do if eigenvalue can not be determined to
the given accuracy with the required max iteration. Is it still possible to extract some
approximated value, even not at the desired accuracy?

- SF_STAT: added interface to evaluation and use of Probability Distribution Function.
PDF can be constructed using gaussian accumulation with Silverman's rule of thumb for the
variance. PDF is then used to evaluate different moments.