4.5.10 UPDATE & BUG-FIX CMAKE
- CMakeLists.txt has been updated.
One bug (spotted by Matteo Secli') has been removed. The bug concerned the
creationg of the new version file in the target installation directory. This
file was created at the execution of cmake, which created a problem if the
target directory was located in a place were write priviliges are not granted
to the user (i.e. it requires sudo to performe make install).
The creation of the file is now part of the installation process, possibly
executed under sudo.

The check of MPI and mpif.h header is now moved above in the script.