4.5.11 UPDATE CMAKE: Problem with mpif.h fixed forever
Well I am a dumb sometime.
The problem with mpif.h in parpack is this.
Parpack includes mpif.h and mpif.h changes among MPI-distributions.
First I was using mpich native mpif.h. This created a misbehavior
when using openmpi (of course). We took time to discover this problem.
The quick fix was manual.

Last week I used CMAKE find-package MPI tools to copy the mpif.h headers
in the proper directories in parpack.

This looked like a fix, but it was not.
>> The real fix is the easiest one: REMOVE THE HEADER FROM PARPACK AND USE
THE -I<path to MPI include> <<
The mpif90 wrapper contains a nice -I path that automatically offers the correct
mpif.h, unless ( as the silly me did) place a local copy of the mpif.h header in the
directory.
Removed the mpif.h everything looks like fine. WTF!