Update
This version of SciFortran removed default support for Lapack/Blas.
We introduced a variable *WITH_BLAS_LAPACK*, which is default OFF/False
to control this feature.

If OFF blas & lapack are assumed to be found
in the operative system. This allows to use highly optimized libraries,
such vecLib in OSX or MKL in Linux.
The user will be required to link against blas/lapack at compilation time
when using scifor. This is easily done in CMake or Make.
We will export this to all CMake in all the codes based on SciFor soon.

If ON blas & lapack are compiled inside scifor (old behavior), then the
user might skip the linking part against blas/lapack when compiling his/her
code. If not, i.e. the user specifies system version of blas/lapack, the
linking will superseed the scifor one.

There is a known problem involving the use of MKL on Linux machines
which requires the compilation flag -ff2c. This flag however shown
to be harmful in OSX at linking stage. The problem is that once
this flag is used all the codes depending on libraries or other
codes using this flag MUST compile with this flag too.
Action is taken in updating CMake files.

More tests are to come for MKL support on Linux and OSX machines.
So far OSX seems to not be able to recognize the presence of MKL.
Too bad,  because VecLib versions are anyway very fast or possibly
faster than MKL. So we might decide to enable MKL only for Linux
machines (that are 99.9% of HPC machines).