Update + Blas/Lapack dependency fixed
This version features a novel support for third-party Blas/Lapack
libraries, namaley OS-provided or MKL.

This is achieved through a suitable setup of the pkg-config file
etc/scifor.pc to be installed. This file is read from the system
when scifor should be used. In this way the user of scifor will
know which libraries to link, e.g.
-lscifor
-lblas -llapack -lscifor
-lmkl_... -lscifor

If no 3rd-party library is found internal BLas/Lapack will be
compiled and used instead. This however does not guarantee the
highest performance on a given machine.

Inside CMake we proceed as follow:
test quietly for MKL
if no :
test quietly for Blas/Lapack
if no :
fall back on internal libraries.

Note that in any code based on CMake compilation that intend to
use Scifor an addendum like the following must be added:
IF("${SCIFOR_LDFLAGS}" MATCHES "blas" OR "lapack")
  FIND_PACKAGE(BLAS REQUIRED)
  FIND_PACKAGE(LAPACK REQUIRED)
ELSEIF("${SCIFOR_LDFLAGS}" MATCHES "mkl")
  FIND_PACKAGE(MKL REQUIRED)
ENDIF()
INCLUDE_DIRECTORIES(BEFORE ${SCIFOR_INCLUDE_DIRS})

To ensure (just a ultra-safety measure) that
support for the same libraries as used to compile Scifor
is available in the system at compilation+linking time.
Otherwise errors would occur. This avoid having weird
error message the user might be scared of.

>> ACTHUNG <<
We patched the ARPACK/PARPACK code to avoid problems with
zdotc from BLAS (read gfortran manual at -ff2c to learn about it)
All instances of *zdotc* in the form of a function have been
replaced with actual call as subroutine
xyz = zdotc(...) ---> call zdotc(xyz, ...)

Getting ready to merge into master