4.6.0 Updated + New features
+ SF_LINALG now support (a rather partial) interface to Scalapack.
Scalapack is searched in the CMakeLists.txt. If found a pre-processing
variable _SCALAPACK is set.
The actual interface is limited to the symmetric/hermitian eigenproblem
solution: pdsyev/r/x/d + pzheevr/x.
Name of the interface: p_eigh
>> Note that Scalapack has an bug in the hermitian procedure which limits
functioning to pzheevr/x case with UPLO="L". I will communicate this bug
to the developer as soon as possible <<
More interfaces to other important routines will come in the next future.
More testing is required both for CMake + Scalapack interface.
NOTE: Usage of Scalapack require to compile, init and run with MPI.

+ SF_IOTOOLS: read/save_array now support Row-major and Col-major ordering.

+ FindMKL: updated version with various possible bug fixes, which should extend
functionality to Unix/Linux machines. Also support for MKL Scalapack/Blacs, to
be tested.