4.6.12 ARPACK BUGFIX
Spotted a bug in the parallel complex version of the post-processing
routine PZNEUPD. This caused the routine to return, in some cases, the
wrong eigenvalues and eigenvectors, skipping the ground state.
Solved by including a newer, community mantained version of arpack
(https://github.com/opencollab/arpack-ng).

To my knowledge the bug persists in the original arpack code.

This has of course potential ripercussions on all the codes that used
the complex version of arpack (dmft-ed, vca, cdmft-lanc-ed). It could be
useful to rerun tests that gave problems to see if this was indeed the
cause. It would be also useful to rerun tests of the real code to see if
the new version of arpack hasn't spawned other issues.