4.6.14 BUG FIX + PATCH + UPDATE
- Found a bug in Arpack, mostly in OSX/Unix, related to Blas *zdotc* malfunctioning.
We already patched this bug, but it re-appeared with the latest arpack update.
To solve the problem we now introduced a patch (zdotc.patch) which is invoked
directly by Cmake. To be tested in OSX. If no further info appear this can be
considered settled down.

- Update: added debug option to arpack interface, i.e. iverbose=.true. add up
verbosity (5) directly in arpack. Expect a lot of information. A better strategy
to fine tune these info will be implemented in the future.