4.6.20 BUG FIX
- Re-fix of the double sqrt in the MPI Lanczos interfaces in SF_SP_LINALG.
THis was fixed yesterday by LCrippa but I have overwritten on it with my latest
(forced !!) commit.

The evaluation of the norm of the eigenvector was miscalculated in MPI mode.

norm_tmp = v*v and then MPI_Allreduce(norm_tmp,norm) to sum up squares.
Instead we had
norm_tmo = sqrt(v*v) and then MPI_allreduce to sum up sqrt of squares, which as math
dictates can not be done.
I.e. the sqrt should be taken after the MPI_allreduce call.
now fixed.