4.6.99 update, fast-forward and BUG FIX WAIT FURTHER NOTICE BEFORE UPDATE. NEXT STABLE RELEASE WILL BE 4.7.0.
a.Fixed the problem with array rank > 7:
b.Fixed a bug which stopped execution of ARPACK after resize of the block
array maxNCV>Ns (observed many times in ED calculations).

a. is now solved using pre-processing macro:
Support for rank+corank<=15 has been introduced in GCC 8.0.0 as part of
fortran2008 standard.
If compiler is older than this rank<=7 only are included in certaing operations.

b. this bug was caused by a bad resize of the block used in P-Arpack, apparently
some thread do not get synced with the correct resized number for maxNCV, causing themn
to hang forever. Fixed using a MPI_allreduce as suggested by Xuanyu Long.