4.7.3 BUGFIX
In version 4.7.0 the routines accounting for rank-8 arrays were
introduced, under specific precompilation flags, for various functions
including MPI_Bcast.

However, a bug was also introduced in the overloaded functions
broadcasting scalar variables: the variable "data" with intent in was
copied into an internal variable "data_", which was then broadcast.

This created a problem whereby the value for the variable stored in the
master was not broadcast to the other cores, which remained with their
respective values.

This of course rendered the code unusable in any parallel scenario.

With this commit I don't remove the data_ variable yet, but I broadcast the
original input one.

The problem may in fact not be completely solved: the reason this was done was,
apparently, an MPI error message warning that scalar input variable could not be broadcast. I have not yet encountered such a message with openmpi 3 and 4 across different setups.

If such a problem should occur, we would have to think of a definitive
solutions.