Add preliminary support for API documentation generation through FORD
For now we set uncommon UTF-8 characters as [pre]docmark[_alt] symbols,
so to deactivate docstring generation: better nothing than a total mess.

Gradually we'd want to parse and standardize the comments across all the
modules (currently there are many different styles, so whatever markers
we choose, the output is very unclean). There is no need to conform to
the defaults of FORD, but some convention has to be taken (and followed).

Even without docstrings the generated html provides already significant
value: all procedures/types/interfaces are searchable, without the usual
noise that grep or similar tools generates (as calls and references are
picked too), all dependencies / call-graphs / inheritance-graphs / etc.
are generated and nicely displayed.

TO FIX (or at least check):

• ERROR in file 'ioread_control.f90': Unexpected procedure call in sourcefile 'ioread_control.f90':
        call sleep(5)

• WARNING: Can not find include file "mpif.h". [MANY OCCURENCIES]

• ERROR in file 'SF_BLACS.f90': Multiple CONTAINS statements present in module 'SF_BLACS':
        contains

• ERROR in file 'SF_MPI.f90': Multiple CONTAINS statements present in module 'SF_MPI':
        contains

• Warning: Error parsing src/SF_SPECIAL/SF_SPECIAL.f90.
        utf-8

• Warning: Error parsing src/SF_SPECIAL/special_functions.f90.
        utf-8