FORD: better temporary `[pre]docmark[_alt]` symbols
For some reason we keep generating unwanted docstrings, it seems that
both docmark_alt and predocmark_alt are not correctly overriden, for
whatever reason. The new custom symbols do not appear to fix the issue,
but at least are less random. If there is really a bug in FORD, making
unfeasible to change the _alt symbols, we have an annoying restriction.