FORD: improve support for `fpp` preprocessor
As documented in the official wiki^, FORD preprocesses only files with
a given set of extensions, by default. These do not include f90, so we
were parsing un-preprocessed files.

In particular this was the reason for

• ERROR in file 'SF_BLACS.f90':
        Multiple CONTAINS statements present in module 'SF_BLACS':
        contains

• ERROR in file 'SF_MPI.f90':
        Multiple CONTAINS statements present in module 'SF_MPI':
        contains

Here we properly tell FORD that we want .f90 files to be preprocessed,
and the _MPI and _SCALAPACK macros to be appear as defined.
Furthermore, while untested (don't have such a old GCC version at hand),
I believe the preprocessor is able to determine that "__GNUC__ > 8": in
fact those implementations are correctly included in the documented API.
I guess if one happens to run FORD locally, and has an older GCC version
installed, it would discard them, which would be the *correct* behavior.

The only warning remaining unsolved at this point is:

• WARNING: Can not find include file "mpif.h"

which I think we can tolerate (and if not we should report the issue to
FORD, given that it's completely legal to expect the mpifort wrapper to
provide that header, so we are definitely not changing what we do here).

__________________________________________________________________________________________
^https://github.com/Fortran-FOSS-Programmers/ford/wiki/Project-File-Options#fpp_extensions