Add github action to autogenerate FORD docs for latest master
It automatically runs FORD for each push event on master [probably we
want to track git tags instead, I'd come to this later]. In any case,
it can be invoked manually too (for debugging purposes).

This workflow closely resembles what we already do in HoneyTools, so at
some point we might want to setup a unique custom action in a suitable
dedicated repository, to be used in all the ford-enabled projects.
[@SamueleGiuli, would you like to help on this?]