4.10.0 Preliminary support for FORD generation of API docs
For now we set uncommon UTF-8 characters as [pre]docmark[_alt] symbols,
so to (almost)^ deactivate docstring generation.

^ We say almost cause there is some bug / misunderstanding on how to
  override default dockmark_alt and predockmark_alt symbols, so we
  still get some noise.

Gradually we'd want to parse and standardize the comments across all the
modules (currently there are many different styles, so whatever marker
set we choose, the output is very messy). There is no need to conform to
the defaults of FORD, but some convention has to be taken (and followed).

Even without docstrings the generated html provides already significant
value: all procedures/types/interfaces are searchable, without the usual
noise brought by grep or similar tools (as calls and references are not
picked); all dependencies / call-graphs / inheritance-graphs / etc. are
generated and nicely displayed in the html output, which can be easily
uploaded to a github page: a github action for that has been already set
up, see < https://qcmplab.github.io/SciFortran >.

Some well-defined refactoring has been made to the source code, with the
aim to ease the parsing for FORD (and improving the codebase itself).

Namely:

- fixed some non UTF-8 characters within `special_functions.f90`
  > see 455dff5f6e0193fffa31b816836fb8bcf6085607

- made `ioread_control.f90` a proper (external) subroutine instead of
  just an included 'code-snippet'
  > see a2cd2b7aba019ae411a126b75ec92311ecab7f72