4.10.1 Fix link error in SF_IOTOOLS (+enlarge testsuite to cover it)
My bad, this has been introduced in commit a2cd2b7aba019ae411a126b75ec92311ecab7f72.

This is probably the ultimate proof that external procedures are dangerous, so here
I convert ioread_control to a (private) module procedure, which is overall better.

The alternative fix would have been to retouch the local CMakeLists file to include
ioread_control.f90 in the src glob, as currently done (extensively) in SF_OPTIMIZE.
> I believe is better to have everything specified in the source, when possible.

—————————————————————————————————————————————————————————————————————————————————————

Link errors are not caught by the build system (it leaves freedom to the user to have
undefined symbols in the library, to be linked later when building an executable).
Again, my bad for not having tested the IOTOOLS subroutines at runtime. Here I ensure
we do not fall into trouble again by expanding the CI testsuite to cover this module.
A test already existed in the module subfolder, so I move and integrate it in the CI,
with some suitable retouch, mainly using the asserting module and filling the arrays
with many different random values, instead of many copies of the same (via an impure
elemental subroutine).

Finally, I fix what I believe to be a typo in ASSERTING, namely default tolerance set
to 1d-5 for double precision scalars (for arrays is 1d-12, surely more appropriate).

Shouldn't ASSERTING become a first-class module in SciFortran? Probably containing the
assert_shape procedures too. It's too nice to not have it available in the library :)